Refactor workspace management components Update handle file change function to check for files before update
