+++ web/src/pages/info_page.tsx
@@ -1,24 +1,17 @@
 import { useStore } from 'zustand'
 import appStore from '../store'
-import { useEffect, useState, useRef } from 'react'
+import { useEffect, useState } from 'react'
 import axios from 'axios'
 import VersionDetails from '@/models/VersionDetails'
 import { toast } from 'sonner'
-import { LuCalendarClock, LuCode, LuExternalLink, LuGithub, LuGlobe, LuShieldCheck, LuTag, LuDownload, LuUpload } from 'react-icons/lu'
+import { LuCalendarClock, LuCode, LuExternalLink, LuGithub, LuGlobe, LuShieldCheck, LuTag } from 'react-icons/lu'
 import { FaEthereum } from 'react-icons/fa6'
 import versionInfo from '../version-info.json'
 import { AlertTriangleIcon } from 'lucide-react'
-import { Button } from '@/components/ui/button'
-import JSZip from 'jszip'
-import { getGenesisHash, getAquaTreeFileName, ensureDomainUrlHasSSL, isValidUrl } from '../utils/functions'
-import { ApiFileInfo } from '@/models/FileInfo'
-import WorkspaceDialogUI from '@/components/workspace_download_dialog_ui'
-import { isAquaTree } from '../utils/functions'
-import { AquaTree } from 'aqua-js-sdk'
+import WorkspaceManagment from '@/components/workspace/WorkspaceManagment'
 
 const InfoPage = () => {
-      const { backend_url, session, setWorkSpaceDowload } = useStore(appStore) // Removed setOpenDialog as dialog is now local
-      const fileInputRef = useRef<HTMLInputElement>(null)
+      const { backend_url } = useStore(appStore)
 
       // const [isOpen, setIsOpen] = useState(false);
       const [versionDetails, setVersionDetails] = useState<VersionDetails>({
@@ -28,13 +21,7 @@ const InfoPage = () => {
             protocol: '1.2.X',
       })
 
-      // New states for workspace operation progress dialog
-      const [isUploading, setIsUploading] = useState(false)
-      const [isDownloading, setIsDownloading] = useState(false)
-      // const [isOperationOpen, setIsOperationOpen] = useState(false)
-      // const [operationMessage, setOperationMessage] = useState('')
-      const [operationProgress, setOperationProgress] = useState(0)
-      // const [operationFileName, setOperationFileName] = useState('')
+      
 
       const fetchVersionDetails = async () => {
             try {
@@ -58,425 +45,9 @@ const InfoPage = () => {
             }
       }, [backend_url])
 
-      const handleDownloadWorkspace = async () => {
-            if (!session?.nonce) {
-                  toast.error("Please connect your wallet first")
-                  return
-            }
+      
 
-            setIsDownloading(true)
-
-            try {
-                  // setIsOperationOpen(true)
-                  // setOperationMessage("Fetching workspace files...")
-                  // setOperationProgress(0)
-                  // setOperationFileName("")
-
-                  // Fetch list of all files (limit=1000000)
-                  const listResponse = await axios.get(`${backend_url}/explorer_files?limit=1000000`, {
-                        headers: {
-                              'nonce': session.nonce
-                        }
-                  })
-
-                  const allFiles: ApiFileInfo[] = listResponse.data.data
-                  const totalFiles = allFiles.length
-                  const zip = new JSZip()
-                  const nameWithHash: { name: string, hash: string }[] = []
-                  const processedFiles = new Set<string>()
-                  // throw new Error("Debugging download workspace")
-
-                  for (let i = 0; i < totalFiles; i++) {
-                        const fileInfo = allFiles[i]
-                        const aquaTree = fileInfo.aquaTree
-                        if (!aquaTree) continue
-
-                        const genesisHash = getGenesisHash(aquaTree)
-                        if (!genesisHash) continue
-
-                        const fileName = getAquaTreeFileName(aquaTree)
-                        if (!fileName) continue
-
-                        console.log("Downloading file: index ", i)
-
-                        // setOperationMessage(`Processing files... (${i + 1}/${totalFiles})`)
-                        // setOperationProgress(Math.round(((i + 1) / totalFiles) * 100))
-                        // setOperationFileName(fileName)
-
-                        setWorkSpaceDowload({
-
-                              fileIndex: i, totalFiles: totalFiles, fileName: fileName
-                        })
-
-
-
-
-                        // Add aqua tree JSON
-                        const aquaTreeFileName = `${fileName}.aqua.json`
-                        zip.file(aquaTreeFileName, JSON.stringify(aquaTree, null, 2))
-
-                        let exist = nameWithHash.find(nwh => nwh.name === aquaTreeFileName)
-                        if (!exist) {
-                              nameWithHash.push({ name: aquaTreeFileName, hash: genesisHash })
-                        }
-
-                        // Process associated files
-                        for (const fileObj of fileInfo.fileObject) {
-                              if (fileObj.fileName && !processedFiles.has(fileObj.fileName)) {
-                                    try {
-                                          let fileContent: ArrayBuffer | string | null | Uint8Array<ArrayBufferLike> = null
-
-                                          if (typeof fileObj.fileContent === 'string' && isValidUrl(fileObj.fileContent)) {
-                                                const actualUrl = ensureDomainUrlHasSSL(fileObj.fileContent)
-                                                // Fetch file content
-                                                const fileResponse = await axios.get(actualUrl, {
-                                                      responseType: 'arraybuffer',
-                                                      headers: { 'nonce': session.nonce }
-                                                })
-                                                fileContent = fileResponse.data
-                                          } else if (typeof fileObj.fileContent === 'string') {
-
-                                                let json = fileObj.fileContent;
-                                                try {
-                                                      const jsonData = JSON.parse(json);
-                                                      fileContent = JSON.stringify(jsonData, null, 2)
-                                                } catch (e) {
-
-                                                      fileContent = fileObj.fileContent
-                                                }
-                                          } else if (fileObj.fileContent instanceof Uint8Array) {
-                                                fileContent = fileObj.fileContent
-                                          } else if (isAquaTree(fileObj.fileContent)) {
-
-                                                let aquatreeData: AquaTree
-                                                if (typeof fileObj.fileContent === 'string') {
-                                                      aquatreeData = JSON.parse(fileObj.fileContent as string)
-                                                } else {
-                                                      aquatreeData = fileObj.fileContent as AquaTree
-                                                }
-                                                try {
-                                                      let genesisHash = getGenesisHash(aquatreeData)
-                                                      let exist = nameWithHash.find(nwh => nwh.name === fileObj.fileName)
-                                                      if (!exist) {
-
-                                                            nameWithHash.push({ name: fileObj.fileName, hash: genesisHash ?? "" })
-                                                      }
-                                                } catch (err) {
-                                                      console.error("Error adding to nameWithHash:", err)
-                                                }
-                                                fileContent = JSON.stringify(fileObj.fileContent, null, 2)
-                                          } else {
-                                                throw new Error("Unsupported file content type")
-                                          }
-
-                                          if (fileContent) {
-
-                                                zip.file(fileObj.fileName, fileContent)
-                                                processedFiles.add(fileObj.fileName)
-                                          }
-                                    } catch (err) {
-                                          console.error(`Failed to download asset ${fileObj.fileName}:`, err)
-                                    }
-                              }
-                        }
-                  }
-
-                  // Create manifest
-                  const aquaManifest = {
-                        type: "aqua_workspace_backup",
-                        version: "1.0.0",
-                        createdAt: new Date().toISOString(),
-                        genesis: "0000000000000000000000000000000000000000000000000000000000000000",
-                        name_with_hash: nameWithHash
-                  }
-                  zip.file("aqua.json", JSON.stringify(aquaManifest, null, 2))
-
-                  // setOperationMessage("Generating ZIP file...")
-
-
-                  const content = await zip.generateAsync({ type: "blob" })
-
-                  const url = window.URL.createObjectURL(content)
-                  const link = document.createElement('a')
-                  link.href = url
-                  link.setAttribute('download', `workspace_${session.address || 'backup'}.zip`)
-                  document.body.appendChild(link)
-                  link.click()
-                  link.parentNode?.removeChild(link)
-                  toast.success("Workspace downloaded successfully")
-
-            } catch (error) {
-                  console.error(error)
-                  toast.error("Failed to download workspace")
-            } finally {
-                  setIsDownloading(false)
-            }
-      }
-
-      const handleUploadWorkspaceClick = () => {
-            fileInputRef.current?.click()
-      }
-
-      const handleFileChange = async (event: React.ChangeEvent<HTMLInputElement>) => {
-            const file = event.target.files?.[0]
-            if (!file) return
-
-            if (!session?.nonce) {
-                  toast.error("Please connect your wallet first")
-                  return
-            }
-
-            // Validate that the file is a zip
-            if (!file.name.toLowerCase().endsWith('.zip')) {
-                  toast.error("Please select a ZIP file")
-                  if (fileInputRef.current) {
-                        fileInputRef.current.value = ''
-                  }
-                  return
-            }
-
-            try {
-                  // Load and validate zip contents
-                  const zip = new JSZip()
-                  const zipContent = await zip.loadAsync(file)
-
-                  // Helper function to decode byte-encoded filenames
-                  const decodeFileName = (name: string): string => {
-                        // Check if the name looks like a comma-separated byte array
-                        if (name.includes(',') && /^[\d,]+$/.test(name)) {
-                              try {
-                                    const bytes = name.split(',').map(b => parseInt(b.trim(), 10))
-                                    return String.fromCharCode(...bytes)
-                              } catch (e) {
-                                    return name
-                              }
-                        }
-                        return name
-                  }
-
-                  // Get all files and decode their names
-                  const allFiles = Object.keys(zipContent.files)
-                  const decodedFileMap = new Map<string, string>() // decoded name -> original key
-
-                  allFiles.forEach(originalKey => {
-                        const decodedName = decodeFileName(originalKey)
-                        decodedFileMap.set(decodedName, originalKey)
-                  })
-
-                  console.log("Decoded files in ZIP:", Array.from(decodedFileMap.keys()))
-                  console.log("Decoded:", decodedFileMap)
-
-                  // Check if aqua.json exists (try multiple variations)
-                  let aquaJsonFile = null
-                  let aquaJsonOriginalKey = null
-
-                  // First try direct access using decoded name
-                  if (decodedFileMap.get("aqua.json")) {
-                        aquaJsonOriginalKey = decodedFileMap.get("aqua.json")!
-                        aquaJsonFile = zipContent.file(aquaJsonOriginalKey)
-                        console.log("Found aqua.json with original key:", aquaJsonOriginalKey)
-                  }
-
-                  if (!aquaJsonFile) {
-                        toast.error("Invalid workspace: aqua.json not found in ZIP")
-                        if (fileInputRef.current) {
-                              fileInputRef.current.value = ''
-                        }
-                        return
-                  }
-
-                  // Read and parse aqua.json
-                  const aquaJsonContent = await aquaJsonFile.async("string")
-                  console.log("aqua.json content (first 200 chars):", aquaJsonContent.substring(0, 200))
-                  console.log("aqua.json content length:", aquaJsonContent.length)
-
-                  let aquaManifest: {
-                        type: string
-                        version: string
-                        createdAt: string
-                        genesis: string
-                        name_with_hash: { name: string, hash: string }[]
-                  }
-
-                  try {
-                        // Check if content might be byte-encoded
-                        let jsonContent = aquaJsonContent
-                        if (aquaJsonContent.includes(',') && /^[\d,\s]+$/.test(aquaJsonContent.substring(0, 100))) {
-                              console.log("Detected byte-encoded JSON content, decoding...")
-                              const bytes = aquaJsonContent.split(',').map(b => parseInt(b.trim(), 10))
-                              jsonContent = String.fromCharCode(...bytes)
-                              console.log("Decoded JSON content (first 200 chars):", jsonContent.substring(0, 200))
-                        }
-
-                        aquaManifest = JSON.parse(jsonContent)
-                  } catch (parseError) {
-                        console.error("JSON parse error:", parseError)
-                        toast.error("Invalid workspace: aqua.json is not valid JSON")
-                        if (fileInputRef.current) {
-                              fileInputRef.current.value = ''
-                        }
-                        return
-                  }
-
-                  // Validate aqua.json structure
-                  if (!aquaManifest.name_with_hash || !Array.isArray(aquaManifest.name_with_hash)) {
-                        toast.error("Invalid workspace: name_with_hash property not found or invalid")
-                        if (fileInputRef.current) {
-                              fileInputRef.current.value = ''
-                        }
-                        return
-                  }
-
-
-
-
-                  //todo loop through all files that end with .aqua.json
-                  // parse them and check if the files in file index exist in the zip
-                  for (const [decodedName, originalKey] of decodedFileMap.entries()) {
-                        if (decodedName.toLowerCase().endsWith('.aqua.json') && decodedName !== 'aqua.json') {
-                              try {
-                                    const aquaTreeFile = zipContent.file(originalKey)
-                                    if (!aquaTreeFile) continue
-                                    const aquaTreeContent = await aquaTreeFile.async("string")
-
-                                    let aquaTreeJsonContent = aquaTreeContent
-                                    if (aquaTreeContent.includes(',') && /^[\d,\s]+$/.test(aquaTreeContent.substring(0, 100))) {
-                                          console.log(`Detected byte-encoded AquaTree JSON content in ${decodedName}, decoding...`)
-                                          const bytes = aquaTreeContent.split(',').map(b => parseInt(b.trim(), 10))
-                                          aquaTreeJsonContent = String.fromCharCode(...bytes)
-                                          console.log(`Decoded AquaTree JSON content in ${decodedName} (first 200 chars):`, aquaTreeJsonContent.substring(0, 200))
-                                    }
-                                    const aquaTree = JSON.parse(aquaTreeJsonContent)
-                                    if (aquaTree.fileIndex && Array.isArray(aquaTree.fileIndex)) {
-                                          for (const fileEntry of aquaTree.fileIndex) {
-                                                const fileName = fileEntry.fileName
-                                                let fileFound = false
-                                                // First try direct access
-                                                if (zipContent.file(fileName)) {
-                                                      fileFound = true
-                                                }
-                                                // If not found, search in decoded names
-                                                if (!fileFound) {
-                                                      const foundDecodedName = Array.from(decodedFileMap.keys()).find(name =>
-                                                            name === fileName ||
-                                                            (name.toLowerCase().endsWith(fileName.toLowerCase()) && !zipContent.files[decodedFileMap.get(name)!].dir)
-                                                      )
-                                                      if (foundDecodedName) {
-                                                            fileFound = true
-                                                      }
-                                                }
-
-                                                if (!fileFound) {
-                                                      toast.error(`Invalid workspace: File ${fileName} listed in ${decodedName} is missing from ZIP`)
-                                                      if (fileInputRef.current) {
-                                                            fileInputRef.current.value = ''
-                                                      }
-                                                      return
-                                                }
-                                          }
-                                    }
-                              } catch (err) {
-                                    console.error(`Error validating AquaTree file ${decodedName}:`, err)
-                                    toast.error(`Invalid workspace: Failed to validate ${decodedName}`)
-                              }
-                        }
-                  }
-                  // Check if all files in name_with_hash exist in the zip
-                  const missingFiles: string[] = []
-                  const decodedNames = Array.from(decodedFileMap.keys())
-
-                  for (const fileEntry of aquaManifest.name_with_hash) {
-                        const fileName = fileEntry.name
-                        let fileFound = false
-
-                        // First try direct access
-                        if (zipContent.file(fileName)) {
-                              fileFound = true
-                        }
-
-                        // If not found, search in decoded names
-                        if (!fileFound) {
-                              const foundDecodedName = decodedNames.find(name =>
-                                    name === fileName ||
-                                    (name.toLowerCase().endsWith(fileName.toLowerCase()) && !zipContent.files[decodedFileMap.get(name)!].dir)
-                              )
-                              if (foundDecodedName) {
-                                    fileFound = true
-                              }
-                        }
-
-                        if (!fileFound) {
-                              missingFiles.push(fileName)
-                        }
-                  }
-
-                  if (missingFiles.length > 0) {
-                        toast.error(`Invalid workspace: ${missingFiles.length} file(s) missing from ZIP. First missing: ${missingFiles[0]}`)
-                        if (fileInputRef.current) {
-                              fileInputRef.current.value = ''
-                        }
-                        return
-                  }
-
-                  toast.success("Workspace validated successfully")
-            } catch (error) {
-                  console.error("Error validating workspace:", error)
-                  toast.error("Failed to validate workspace ZIP file")
-                  if (fileInputRef.current) {
-                        fileInputRef.current.value = ''
-                  }
-                  return
-            }
-
-            setIsUploading(true)
-            setOperationProgress(0)
-
-            const formData = new FormData()
-            formData.append('file', file)
-
-            let isSuccess = false
-            try {
-                  // setIsOperationOpen(true)
-                  // setOperationMessage("Uploading workspace...")
-                  // setOperationProgress(0)
-                  // setOperationFileName(file.name) // Set filename for display
-
-                  const response = await axios.post(`${backend_url}/explorer_workspace_upload`, formData, {
-                        headers: {
-                              'nonce': session.nonce,
-                              'Content-Type': 'multipart/form-data'
-                        },
-                        onUploadProgress: (progressEvent) => {
-                              if (progressEvent.total) {
-                                    const percentCompleted = Math.round((progressEvent.loaded * 100) / progressEvent.total)
-                                    setOperationProgress(percentCompleted)
-                              }
-                        }
-                  })
-
-                  if (response.status === 200) {
-
-                        isSuccess = true
-                  }
-            } catch (error) {
-                  console.error(error)
-                  toast.error("Failed to upload workspace")
-            } finally {
-
-                  // setIsOperationOpen(false)
-                  // setOperationProgress(0)
-                  // setOperationFileName('')
-                  if (fileInputRef.current) {
-                        fileInputRef.current.value = ''
-                  }
-                  if (isSuccess) {
-                        toast.success("Workspace uploaded successfully")
-                        window.location.href = "/app"
-                  } else {
-                        setIsUploading(false)
-                  }
-            }
-      }
+     
 
       return (
             <div className="container mx-auto py-3 px-2 sm:px-4">
@@ -486,42 +57,7 @@ const InfoPage = () => {
                               {/* Main content grid */}
                               <div className="grid grid-cols-12 md:grid-cols-6 gap-4 sm:gap-6">
                                     {/* Workspace Management Card */}
-                                    <div className="col-span-12 md:col-span-6 bg-white dark:bg-gray-800 rounded-xl shadow-sm border border-gray-200 dark:border-gray-700 overflow-hidden">
-                                          <div className="border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800/50 px-3 sm:px-6 py-3 sm:py-4">
-                                                <h2 className="text-lg sm:text-xl font-semibold flex items-center gap-2">
-                                                      <LuCode className="h-5 w-5 text-primary" />
-                                                      Workspace Management
-                                                </h2>
-                                          </div>
-                                          <div className="p-3 sm:p-6 flex flex-wrap gap-4">
-                                                <Button
-                                                      onClick={handleDownloadWorkspace}
-                                                      disabled={isDownloading}
-                                                      className="flex items-center gap-2"
-                                                >
-                                                      <LuDownload className="h-4 w-4" />
-                                                      {isDownloading ? "Downloading..." : "Download Workspace"}
-                                                </Button>
-
-                                                <input
-                                                      type="file"
-                                                      ref={fileInputRef}
-                                                      onChange={handleFileChange}
-                                                      accept=".zip"
-                                                      className="hidden"
-                                                />
-
-                                                <Button
-                                                      onClick={handleUploadWorkspaceClick}
-                                                      disabled={isUploading}
-                                                      variant="outline"
-                                                      className="flex items-center gap-2"
-                                                >
-                                                      <LuUpload className="h-4 w-4" />
-                                                      {isUploading ? "Uploading..." : "Upload Workspace"}
-                                                </Button>
-                                          </div>
-                                    </div>
+                                    <WorkspaceManagment />
 
                                     {/* Version info card */}
                                     <div className="col-span-12 md:col-span-4 bg-white dark:bg-gray-800 rounded-xl shadow-sm border border-gray-200 dark:border-gray-700 overflow-hidden">
@@ -694,24 +230,6 @@ const InfoPage = () => {
                         </div>
                   </div>
 
-                  /* Workspace Operation Dialog */
-                  {(isUploading || isDownloading) && (
-                        <div className="fixed inset-0 bg-white/30 dark:bg-black/30 backdrop-blur-sm flex items-center justify-center z-50">
-                              <div className="bg-white dark:bg-gray-800 rounded-lg shadow-lg p-6 w-11/12 max-w-md">
-                                    <WorkspaceDialogUI
-                                          title={isUploading ? "Uploading Workspace" : "Downloading Workspace"}
-                                          uploadProgress={operationProgress}
-                                          isUploading={isUploading}
-                                          isDone={() => {
-                                                setIsUploading(false)
-                                                setIsDownloading(false)
-                                                setOperationProgress(0)
-                                          }
-                                          }
-                                    />
-                              </div>
-                        </div>
-                  )}
             </div>
       )
 }
